+++ index.d.ts
@@ -29,8 +29,9 @@ declare module "react-native-appsflyer" {
 
     export interface InitSDKOptions {
         devKey:string;
-        appId?:string; //for iOS
+        appId?:string; // iOS only
         isDebug?:boolean;
+        timeToWaitForAdvertiserID?:number; // iOS only
     }
 
     export interface SetEmailsOptions {
@@ -51,44 +52,40 @@ declare module "react-native-appsflyer" {
         onInstallConversionFailure(callback:(data:ConversionData)=>any): () => void;
         onAppOpenAttribution(callback:(data:any)=>any): () => void;
 
-        initSdk(options:InitSDKOptions, successC?:SuccessCB, errorC?:ErrorCB): Response<string>
-        trackEvent(eventName:string, eventValues:object, successC?:SuccessCB, errorC?:ErrorCB): Response<string>
-        setUserEmails(options:SetEmailsOptions, successC?:SuccessCB, errorC?:ErrorCB): void
+        initSdk(options: InitSDKOptions, successC?: SuccessCB, errorC?: ErrorCB): Response<string>
+        logEvent(eventName: string, eventValues: object, successC?: SuccessCB, errorC?: ErrorCB): Response<string>
+        setUserEmails(options: SetEmailsOptions, successC?: SuccessCB, errorC?: ErrorCB): void
         setAdditionalData(additionalData: object, successC?: SuccessCB): void
         getAppsFlyerUID(callback: (error: Error, uid: string) => any): void
         setCustomerUserId(userId: string, successC?: SuccessCB): void
-        stopTracking(isStopTracking: boolean, successC?: SuccessCB): void
+        stop(isStopped: boolean, successC?: SuccessCB): void
         setAppInviteOneLinkID(oneLinkID: string, successC?: SuccessCB): void
         generateInviteLink(params: GenerateInviteLinkParams, successC?: SuccessCB, errorC?: ErrorCB): void
-        trackCrossPromotionImpression(appId: string, campaign: string): void
-        trackAndOpenStore(appId: string, campaign: string, params: object): void
+        logCrossPromotionImpression(appId: string, campaign: string, parameters: object): void
+        logCrossPromotionAndOpenStore(appId: string, campaign: string, params: object): void
         setCurrencyCode(currencyCode: string, successC: SuccessCB): void
-        setDeviceTrackingDisabled(isDeviceTrackingDisabled: boolean, successC: SuccessCB): void
+        anonymizeUser(shouldAnonymize: boolean, successC: SuccessCB): void
         setOneLinkCustomDomains(domains: string[], successC?: SuccessCB, errorC?: ErrorCB): void
         setResolveDeepLinkURLs(urls: string[], successC?: SuccessCB, errorC?: ErrorCB): void
         performOnAppAttribution(urlString, callback): void
         setSharingFilterForAllPartners(): void
         setSharingFilter(partners, successC, errorC): void
+        logLocation(longitude: number, latitude: number, callback: SuccessCB): void
 
         /**
          * For iOS Only
          * */
         trackAppLaunch(): void
         trackLocation(longitude: number, latitude: number, callback: SuccessCB): void
-        disableAdvertiserIdentifier(shouldDisable: boolean):void
+        disableAdvertisingIdentifier(shouldDisable: boolean):void
         disableCollectASA(shouldDisable: boolean):void
 
         /**
          * For Android Only
          * */
-        sendDeepLinkData(callback:any): void
-        setCollectIMEI(isCollect:boolean, successC?:SuccessCB): void
-        setCollectAndroidID(isCollect:boolean, successC?:SuccessCB): void
-        updateServerUninstallToken(token:string, successC?:SuccessCB): void
-        /** @deprecated - please use updateServerUninstallToken **/
-        setGCMProjectNumber(gcmProjectNumber:any, successC?:SuccessCB, errorC?:ErrorCB): void
-        /** @deprecated - please use updateServerUninstallToken **/
-        enableUninstallTracking(gcmProjectNumber:any, successC?:SuccessCB): void
+        setCollectIMEI(isCollect: boolean, successC?: SuccessCB): void
+        setCollectAndroidID(isCollect: boolean, successC?: SuccessCB): void
+        updateServerUninstallToken(token: string, successC?: SuccessCB): void
     };
 
     export default appsFlyer;